export things properly in react-email
