+++ packages/react-email/tsdown.config.ts
@@ -1,8 +1,16 @@
 import { defineConfig } from 'tsdown';
 
-export default defineConfig({
-  dts: false,
-  entry: ['./src/cli/index.ts'],
-  format: ['esm'],
-  outDir: 'dist/cli',
-});
+export default defineConfig([
+  {
+    dts: false,
+    entry: ['./src/cli/index.ts'],
+    format: ['esm'],
+    outDir: 'dist/cli',
+  },
+  {
+    dts: true,
+    entry: ['./src/index.ts'],
+    format: ['esm', 'cjs'],
+    outDir: 'dist',
+  },
+]);
